+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT66991YBTXA9SV0F0V7BGEM"
+        "01KT6ZVW2CWXP12JSTT1KPEAZS"
     )
     r"""The unique identifier of the data source"""