+++ packages/orq-rc/RELEASES.md
@@ -8429,3 +8429,13 @@ Based on:
 - [python v4.10.0-rc.75] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.75] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.75 - packages/orq-rc
+
+## 2026-06-03 15:03:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.81] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.81] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.81 - packages/orq-rc